chore: updates per pr review
