+++ crates/net/network/src/lib.rs
@@ -144,6 +144,7 @@ mod network;
 mod session;
 mod state;
 mod swarm;
+mod trusted_peers_resolver;
 
 pub use reth_eth_wire::{DisconnectReason, HelloMessageWithProtocols};
 pub use reth_eth_wire_types::{EthNetworkPrimitives, NetworkPrimitives};
