+++ crates/net/network/src/peers.rs
@@ -4,14 +4,15 @@ use crate::{
     error::SessionError,
     session::{Direction, PendingSessionHandshakeError},
     swarm::NetworkConnectionState,
+    trusted_peers_resolver::TrustedPeersResolver,
 };
-use futures::{future::BoxFuture, stream::FuturesUnordered, FutureExt, StreamExt};
+use futures::StreamExt;
 
 use reth_eth_wire::{errors::EthStreamError, DisconnectReason};
 use reth_ethereum_forks::ForkId;
 use reth_net_banlist::BanList;
 use reth_network_api::test_utils::{PeerCommand, PeersHandle};
-use reth_network_peers::{NodeRecord, PeerId, TrustedPeer};
+use reth_network_peers::{NodeRecord, PeerId};
 use reth_network_types::{
     peers::{
         config::PeerBackoffDurations,
@@ -35,57 +36,6 @@ use tokio::{
 };
 use tokio_stream::wrappers::UnboundedReceiverStream;
 use tracing::{trace, warn};
-
-/// `TrustedPeersResolver` periodically spawns DNS resolution tasks for trusted peers.
-/// It returns a resolved (`PeerId`, `NodeRecord`) update when one of its in‑flight tasks completes.
-#[derive(Debug)]
-pub struct TrustedPeersResolver {
-    /// The timer that triggers a new resolution cycle.
-    pub trusted_peers: Vec<TrustedPeer>,
-    /// The timer that triggers a new resolution cycle.
-    pub interval: Interval,
-    /// Futures for currently in‑flight resolution tasks.
-    pub pending: FuturesUnordered<BoxFuture<'static, (PeerId, Result<NodeRecord, io::Error>)>>,
-}
-
-impl TrustedPeersResolver {
-    /// Create a new resolver with the given trusted peers and resolution interval.
-    pub fn new(trusted_peers: Vec<TrustedPeer>, resolve_interval: Interval) -> Self {
-        Self { trusted_peers, interval: resolve_interval, pending: FuturesUnordered::new() }
-    }
-
-    /// Update the resolution interval (useful for testing purposes)
-    pub fn set_interval(&mut self, interval: Interval) {
-        self.interval = interval;
-    }
-
-    /// Poll the resolver.
-    /// When the interval ticks, new resolution futures for each trusted peer are spawned.
-    /// If a future completes successfully, it returns the resolved (`PeerId`, `NodeRecord`).
-    pub fn poll(&mut self, cx: &mut Context<'_>) -> Poll<Option<(PeerId, NodeRecord)>> {
-        if self.pending.is_empty() && self.interval.poll_tick(cx).is_ready() {
-            for trusted in self.trusted_peers.iter().cloned() {
-                let peer_id = trusted.id;
-                let task = async move {
-                    let result = trusted.resolve().await;
-                    (peer_id, result)
-                }
-                .boxed();
-                self.pending.push(task);
-            }
-        }
-
-        match self.pending.poll_next_unpin(cx) {
-            Poll::Ready(Some((peer_id, Ok(record)))) => Poll::Ready(Some((peer_id, record))),
-            Poll::Ready(Some((peer_id, Err(e)))) => {
-                warn!(target: "net::peers", "Failed to resolve trusted peer {:?}: {:?}", peer_id, e);
-                Poll::Pending
-            }
-            Poll::Ready(None) => Poll::Ready(None),
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
 
 /// Maintains the state of _all_ the peers known to the network.
 ///
@@ -977,6 +927,21 @@ impl PeersManager {
         }
     }
 
+    fn on_resolved_peer(&mut self, peer_id: PeerId, new_record: NodeRecord) {
+        if let Some(peer) = self.peers.get_mut(&peer_id) {
+            let new_addr = PeerAddr::new_with_ports(
+                new_record.address,
+                new_record.tcp_port,
+                Some(new_record.udp_port),
+            );
+
+            if peer.addr != new_addr {
+                peer.addr = new_addr;
+                trace!(target: "net::peers", ?peer_id, "Updated trusted peer address via DNS resolution");
+            }
+        }
+    }
+
     /// Keeps track of network state changes.
     pub fn on_network_state_change(&mut self, state: NetworkConnectionState) {
         self.net_connection_state = state;
@@ -1050,17 +1015,7 @@ impl PeersManager {
             }
 
             if let Poll::Ready(Some((peer_id, new_record))) = self.trusted_peers_resolver.poll(cx) {
-                if let Some(peer) = self.peers.get_mut(&peer_id) {
-                    let new_addr = PeerAddr::new_with_ports(
-                        new_record.address,
-                        new_record.tcp_port,
-                        Some(new_record.udp_port),
-                    );
-                    if peer.addr != new_addr {
-                        peer.addr = new_addr;
-                        trace!(target: "net::peers", ?peer_id, "Updated trusted peer address via DNS resolution");
-                    }
-                }
+                self.on_resolved_peer(peer_id, new_record);
             }
 
             if self.queued_actions.is_empty() {
